Work around reserved area in flash memory
There is a section of reserved memory in flash in the range
0x40000-0x110000.  When we read or write a codeplug larger than
256K (0x40000) bytes, we must skip this region.

The upper bound address for codeplugs is 0x200000 (2048K).
We translate any addresses greater than 256K that would yield a
new address of less than 2048K.

If the flash address is between 256K (inclusive) and (2048-832)K
(exclusive), we add 832K to the address before doing the operation.

Similarly for erasing blocks of flash.